Factor logic out of ServerService into Server
